SSA-018: Update project structure to allow for 'v2' implementation of sorts. Update dependencies
