+++ app/sorts/v1/BucketSort.scala
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package v1.sorts
+package sorts.v1
 
+import sorts.common.Sort
 import scala.annotation.tailrec
 
 object BucketSort extends Sort[BigDecimal] {
@@ -24,7 +25,7 @@ object BucketSort extends Sort[BigDecimal] {
            - Sort items upon insertion into a bucket rather than the whole bucket at the end
    */
 
-  override protected[sorts] def sort(nums: Seq[BigDecimal]): Seq[BigDecimal] = {
+  protected[sorts] def sort(nums: Seq[BigDecimal]): Seq[BigDecimal] = {
 
     @tailrec
     def doSort(nums: Seq[BigDecimal], buckets: Seq[Seq[BigDecimal]] = Nil): Seq[BigDecimal] = {
