+++ app/sorts/v1/CountingSort.scala
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package v1.sorts
+package sorts.v1
 
+import sorts.common.Sort
 import scala.annotation.tailrec
 
 object CountingSort extends Sort[Integer] {
