+++ app/sorts/v1/ExchangeSort.scala
@@ -1,4 +1,7 @@
-package v1.sorts
+package sorts.v1
+
+import sorts.common.Sort
+import Sort.swap
 
 import scala.annotation.tailrec
