+++ app/sorts/v1/MergeSort.scala
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package v1.sorts
+package sorts.v1
 
+import sorts.common.Sort
 import scala.annotation.tailrec
 
 object MergeSort extends Sort[BigDecimal] {
-  override protected[sorts] def sort(nums: Seq[BigDecimal]): Seq[BigDecimal] = doSort(nums)
+  protected[sorts] def sort(nums: Seq[BigDecimal]): Seq[BigDecimal] = doSort(nums)
 
   def doSort(nums: Seq[BigDecimal]): Seq[BigDecimal] = {
     @tailrec
