+++ test/sorts/v1/BubbleSortSpec.scala
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package v1.sorts
+package sorts.v1
 
 import support.SortingSpec
 
-class HeapSortSpec extends SortingSpec {
-  sortingSmokeTest(HeapSort)
+class BubbleSortSpec extends SortingSpec {
+  sortingSmokeTest(BubbleSort)
 }
