+++ test/sorts/v1/MergeSortSpec.scala
@@ -14,19 +14,10 @@
  * limitations under the License.
  */
 
-package v1.sorts
+package sorts.v1
 
-trait Sort[NumType <: Number] {
-  protected[sorts] def sort(nums: Seq[NumType]): Seq[NumType]
+import support.SortingSpec
 
-  def filteredSort(nums: Seq[NumType]): Seq[NumType] = nums match {
-    case Nil => Nil
-    case _ :: Nil => nums
-    case _ => sort(nums)
-  }
-
-  protected[sorts] def swap(nums: Seq[NumType], index1: Int, index2: Int): Seq[NumType] = {
-    if (index1 == index2) nums
-    else nums.slice(0, index1) ++ Seq(nums(index2)) ++nums.slice(index1 + 1, index2) ++ Seq(nums(index1)) ++ nums.drop(index2 + 1)
-  }
+class MergeSortSpec extends SortingSpec {
+  sortingSmokeTest(MergeSort)
 }
