+++ test/sorts/v1/QuickSortSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v1.sorts
+package sorts.v1
 
 import support.SortingSpec
